Add adequate JSONObject handling to pass the wanted parameter to DockerService, add JSONObject to code-execution's pom.xml, moved GitConfigProperties to separate own folder
