+++ code-execution/pom.xml
@@ -13,6 +13,7 @@
 
     <properties>
         <docker-java.version>3.2.7</docker-java.version>
+        <json.version>20180130</json.version>
     </properties>
 
     <dependencies>
@@ -23,6 +24,11 @@
         <dependency>
             <groupId>org.springframework.retry</groupId>
             <artifactId>spring-retry</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.json</groupId>
+            <artifactId>json</artifactId>
+            <version>${json.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
