+++ code-execution/src/main/java/com/dojo/codeexecution/config/github/GitConfiguration.java
@@ -1,4 +1,4 @@
-package com.dojo.codeexecution.config;
+package com.dojo.codeexecution.config.github;
 
 import org.kohsuke.github.GitHub;
 import org.kohsuke.github.GitHubBuilder;
