+++ code-execution/src/main/java/com/dojo/codeexecution/controller/GitEventReceiver.java
@@ -1,6 +1,7 @@
 package com.dojo.codeexecution.controller;
 
 import com.dojo.codeexecution.service.DockerService;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,6 +12,8 @@ import java.util.Map;
 
 @RestController
 public class GitEventReceiver {
+    public static final String PAYLOAD_FIELD = "repository";
+    public static final String PAYLOAD_KEY = "name";
 
     @Autowired
     DockerService dockerService;
@@ -40,8 +43,10 @@ public class GitEventReceiver {
 
     //Currently not able to trigger the webhook which calls this endpoint from github
     @GetMapping(path = "/run")
-    public String runContainer(@RequestBody Map<String, String> payload) {
-        dockerService.runContainer(payload.get("full_name"));
+    public String runContainer(@RequestBody Map<String, Object> payload) {
+        Object imageTag = new JSONObject(payload).getJSONObject(PAYLOAD_FIELD)
+                .get(PAYLOAD_KEY);
+        dockerService.runContainer((String) imageTag);
         return "OK";
     }
 }
