+++ code-execution/src/main/resources/application.properties
@@ -8,7 +8,7 @@ git.user=
 git.tokenCreateRepo=
 git.parentRepositoryName=
 git.webhookAddress=http://192.168.1.104:8081/
-git.webhookConfig.url=http://192.168.1.104:8081/pushEvent
+git.webhookConfig.url=http://192.168.1.104:8081/run
 git.webhookConfig.content_type=json
 git.webhookConfig.insecure_ssl=0
